Awwad
MY WORK SECTION
I worked on the Team Members section. This section has another three small sections inside, Basically I used flexbox in the main section to arrange each member section in one column, and inside each member section I used flexbox too to arrannge the inner information about the member.

ISSUES/CHALLENGES IN MY CODE
In the main section, I wanted to put a background image with low opacity and to look exactly under the session, but didnt find a way to do it with the background image --> so I created another div in the main which contains the wanted background image, I positioned it absolute to the main section with top right bottom left 0, and with small opacity. I used z-index:3 for the main section so we can see the background (actually the div) under the main session.